last edit

+++ AspCore-Conic-Erp-RestApi/Zkt/ZkemClient.cs
@@ -201,7 +201,7 @@ namespace AspCore_Conic_Erp_RestApi
         private  void zkemClient_OnAttTransactionEx(string EnrollNumber, int IsInValid, int AttState, int VerifyMethod, int Year, int Month, int Day, int Hour, int Minute, int Second, int WorkCode)
         {
             DateTime datetime = new DateTime(Year, Month, Day, Hour, Minute, 0);
-            int ID = Convert.ToInt32(EnrollNumber);
+            long ID = Convert.ToInt32(EnrollNumber);
             string TableName = "";
             var member = DB.Members.Where(m => m.Id == ID).FirstOrDefault();
             var Employee = DB.Employees.Where(m => m.Id == ID).FirstOrDefault();
